+++ built_json/test/built_set_serializer_test.dart
@@ -6,49 +6,94 @@ import 'package:built_collection/built_collection.dart';
 import 'package:built_json/built_json.dart';
 import 'package:test/test.dart';
 
+// Note: BuiltSet preserves order, so comparisons in these tests can assume a
+// specific ordering. In fact, these tests are exactly the BuiltListSerializer tests
+// with "list" replaced by "set".
 void main() {
-  final serializers = new Serializers();
-
-  group('BuiltSet with known genericType', () {
+  group('BuiltSet with known specifiedType but missing builder', () {
     final data = new BuiltSet<int>([1, 2, 3]);
-    final genericType =
-        const GenericType(BuiltSet, const [const GenericType(int)]);
+    final specifiedType =
+    const FullType(BuiltSet, const [const FullType(int)]);
+    final serializers = new Serializers();
+    final serialized = [1, 2, 3];
+
+    test('serialize throws', () {
+      expect(() => serializers.serialize(data, specifiedType: specifiedType),
+          throws);
+    });
+
+    test('deserialize throws', () {
+      expect(
+          () =>
+          serializers.deserialize(serialized, specifiedType: specifiedType),
+          throws);
+    });
+  });
+
+  group('BuiltSet with known specifiedType and correct builder', () {
+    final data = new BuiltSet<int>([1, 2, 3]);
+    final specifiedType =
+    const FullType(BuiltSet, const [const FullType(int)]);
+    final serializers = (new Serializers().toBuilder()
+      ..addBuilderFactory(specifiedType, () => new SetBuilder<int>())).build();
     final serialized = [1, 2, 3];
 
     test('can be serialized', () {
-      expect(serializers.serialize(data, genericType: genericType), serialized);
+      expect(serializers.serialize(data, specifiedType: specifiedType),
+          serialized);
     });
 
     test('can be deserialized', () {
-      expect(
-          serializers.deserialize(serialized, genericType: genericType), data);
+      expect(serializers.deserialize(serialized, specifiedType: specifiedType),
+          data);
     });
 
-    test('loses generic type without builder', () {
+    test('keeps generic type when deserialized', () {
       expect(
           serializers
-              .deserialize(serialized, genericType: genericType)
-              .runtimeType
-              .toString(),
-          'BuiltSet<Object>');
-    });
-
-    test('keeps generic type with builder', () {
-      final genericSerializer = (serializers.toBuilder()
-        ..addBuilderFactory(genericType, () => new SetBuilder<int>())).build();
-
-      expect(
-          genericSerializer
-              .deserialize(serialized, genericType: genericType)
+              .deserialize(serialized, specifiedType: specifiedType)
               .runtimeType
               .toString(),
           'BuiltSet<int>');
     });
   });
 
-  group('BuiltSet with unknown genericType', () {
+  group('BuiltSet nested with known specifiedType and correct builders', () {
+    final data = new BuiltSet<BuiltSet<int>>([
+      new BuiltSet<int>([1, 2, 3]),
+      new BuiltSet<int>([4, 5, 6]),
+      new BuiltSet<int>([7, 8, 9])
+    ]);
+    final specifiedType = const FullType(BuiltSet, const [
+      const FullType(BuiltSet, const [const FullType(int)])
+    ]);
+    final serializers = (new Serializers().toBuilder()
+      ..addBuilderFactory(
+          specifiedType, () => new SetBuilder<BuiltSet<int>>())
+      ..addBuilderFactory(
+          const FullType(BuiltSet, const [const FullType(int)]),
+          () => new SetBuilder<int>())).build();
+    final serialized = [
+      [1, 2, 3],
+      [4, 5, 6],
+      [7, 8, 9]
+    ];
+
+    test('can be serialized', () {
+      expect(serializers.serialize(data, specifiedType: specifiedType),
+          serialized);
+    });
+
+    test('can be deserialized', () {
+      expect(serializers.deserialize(serialized, specifiedType: specifiedType),
+          data);
+    });
+  });
+
+  group('BuiltSet with unknown specifiedType and no builders', () {
     final data = new BuiltSet<int>([1, 2, 3]);
-    final genericType = const GenericType();
+    final specifiedType = FullType.unspecified;
+    final serializers = new Serializers();
     final serialized = [
       'set',
       ['int', 1],
@@ -57,12 +102,14 @@ void main() {
     ];
 
     test('can be serialized', () {
-      expect(serializers.serialize(data, genericType: genericType), serialized);
+      expect(serializers.serialize(data, specifiedType: specifiedType),
+          serialized);
     });
 
     test('can be deserialized', () {
-      expect(
-          serializers.deserialize(serialized, genericType: genericType), data);
+      expect(serializers.deserialize(serialized, specifiedType: specifiedType),
+          data);
     });
   });
 }
+