+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.0.3
+
+- Stop falling back to "<Object>" for collection type parameters.
+- Extend code generation to add the builders needed for collections.
+- Differentiate Object from unspecified.
+
 ## 0.0.2
 
 - Rewrite to fully use static type information.